pyproject.toml
[project]
name = "servicenow-to-gbnf"
version = "0.1.0"
description = "ServiceNow → GBNF grammar + iii worker generator. Open-source deterministic AI connector."
authors = [{ name = "intelege-ma" }]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "jinja2>=3.1",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "rich>=13",
]

[project.scripts]
"servicenow-to-gbnf" = "servicenow_to_gbnf.cli:app"

[tool.uv]
dev-dependencies = ["pytest>=8", "ruff", "mypy"]

[tool.ruff]
line-length = 120
